splash25-db/requirements.txt
pandas>=2.0.0
psycopg2-binary>=2.9.0
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0
